Split simulation into simulation and circuit
