+++ vire-api/src/main/kotlin/net/voxelpi/vire/api/simulation/event/component/ComponentConfigureEvent.kt
@@ -1,0 +1,10 @@
+package net.voxelpi.vire.api.simulation.event.component
+
+import net.voxelpi.vire.api.simulation.component.Component
+
+/**
+ * An event that is called when a component is configured.
+ */
+data class ComponentConfigureEvent(
+    override val component: Component,
+) : ComponentEvent
