+++ vire-api/src/main/kotlin/net/voxelpi/vire/api/simulation/event/component/ComponentDestroyEvent.kt
@@ -1,4 +1,4 @@
-package net.voxelpi.vire.api.simulation.event.simulation.component
+package net.voxelpi.vire.api.simulation.event.component
 
 import net.voxelpi.vire.api.simulation.component.Component
