+++ vire-api/src/main/kotlin/net/voxelpi/vire/api/simulation/event/network/node/NetworkNodeCreateEvent.kt
@@ -1,4 +1,4 @@
-package net.voxelpi.vire.api.simulation.event.simulation.network.node
+package net.voxelpi.vire.api.simulation.event.network.node
 
 import net.voxelpi.vire.api.simulation.network.NetworkNode
