+++ vire-engine/src/main/kotlin/net/voxelpi/vire/engine/VireImplementation.kt
@@ -8,7 +8,7 @@ import net.voxelpi.vire.engine.simulation.statemachine.VireStateMachineFactory
 object VireImplementation : Vire {
 
     init {
-        Vire.stateMachineFactory.register(VireStateMachineFactory())
+        Vire.register(this)
     }
 
     override val brand: String = "Vire Engine"
@@ -18,6 +18,8 @@ object VireImplementation : Vire {
 
     override val longVersion: String
         get() = "${VireBuildParameters.VERSION}-${VireBuildParameters.GIT_COMMIT}"
+
+    override val stateMachineFactory: VireStateMachineFactory = VireStateMachineFactory()
 
     override fun createSimulation(libraries: List<Library>): VireSimulation {
         return VireSimulation(libraries)
