+++ vire-engine/src/test/kotlin/net/voxelpi/vire/engine/simulation/VireSimulationTest.kt
@@ -13,76 +13,78 @@ import org.junit.jupiter.api.Test
 class VireSimulationTest {
 
     private lateinit var simulation: VireSimulation
+    private lateinit var circuit: VireCircuit
 
     @BeforeEach
     fun setUp() {
         simulation = VireImplementation.createSimulation(emptyList())
+        circuit = simulation.circuit
     }
 
     @Test
     fun connect2NodesInSameNetwork() {
-        val network = simulation.createNetwork()
+        val network = circuit.createNetwork()
         val node0 = network.createNode(emptyList())
         val node1 = network.createNode(listOf(node0))
         val node2 = network.createNode(listOf(node0))
 
-        assert(!simulation.areNodesConnectedDirectly(node1, node2)) { "Nodes are connected before connecting them." }
-        simulation.createNetworkNodeConnection(node1, node2)
-        assert(simulation.areNodesConnectedDirectly(node1, node2)) { "Nodes are not connected after connecting them." }
+        assert(!circuit.areNodesConnectedDirectly(node1, node2)) { "Nodes are connected before connecting them." }
+        circuit.createNetworkNodeConnection(node1, node2)
+        assert(circuit.areNodesConnectedDirectly(node1, node2)) { "Nodes are not connected after connecting them." }
     }
 
     @Test
     fun connect2NodesInDifferentNetworks() {
-        val network1 = simulation.createNetwork()
+        val network1 = circuit.createNetwork()
         val node1 = network1.createNode(emptyList())
 
-        val network2 = simulation.createNetwork()
+        val network2 = circuit.createNetwork()
         val node2 = network2.createNode(emptyList())
 
-        simulation.createNetworkNodeConnection(node1, node2)
+        circuit.createNetworkNodeConnection(node1, node2)
 
-        assertEquals(1, simulation.networks().size) { "Only one network should exist after the merge." }
-        assert(simulation.areNodesConnectedDirectly(node1, node2)) { "Nodes are not connected after the merge." }
+        assertEquals(1, circuit.networks().size) { "Only one network should exist after the merge." }
+        assert(circuit.areNodesConnectedDirectly(node1, node2)) { "Nodes are not connected after the merge." }
     }
 
     @Test
     fun removeNodeConnectionWithSplit() {
-        val network = simulation.createNetwork()
+        val network = circuit.createNetwork()
         val nodeA = network.createNode(emptyList())
         val nodeB = network.createNode(listOf(nodeA))
 
-        assertEquals(1, simulation.networks().size)
-        assertEquals(2, simulation.networkNodes().size)
+        assertEquals(1, circuit.networks().size)
+        assertEquals(2, circuit.networkNodes().size)
 
-        simulation.removeNetworkNodeConnection(nodeA, nodeB)
+        circuit.removeNetworkNodeConnection(nodeA, nodeB)
 
-        assertEquals(2, simulation.networks().size)
-        assertEquals(2, simulation.networkNodes().size)
+        assertEquals(2, circuit.networks().size)
+        assertEquals(2, circuit.networkNodes().size)
         assert(nodeA.network.uniqueId != nodeB.network.uniqueId)
     }
 
     @Test
     fun removeNodeFromNetworkWithSplit() {
-        val network = simulation.createNetwork()
+        val network = circuit.createNetwork()
         val nodeA = network.createNode(emptyList())
         val nodeMiddle = network.createNode(listOf(nodeA))
         val nodeB = network.createNode(listOf(nodeMiddle))
 
-        assertEquals(1, simulation.networks().size)
-        assertEquals(3, simulation.networkNodes().size)
+        assertEquals(1, circuit.networks().size)
+        assertEquals(3, circuit.networkNodes().size)
 
         nodeMiddle.remove()
 
-        assertEquals(2, simulation.networks().size)
-        assertEquals(2, simulation.networkNodes().size)
+        assertEquals(2, circuit.networks().size)
+        assertEquals(2, circuit.networkNodes().size)
         assert(nodeA.network != nodeB.network)
     }
 
     @Test
     fun clear() {
-        simulation.createNetwork()
-        simulation.clear()
-        assert(simulation.networks().isEmpty()) { "Not all networks have been deleted." }
+        circuit.createNetwork()
+        circuit.clear()
+        assert(circuit.networks().isEmpty()) { "Not all networks have been deleted." }
     }
 
     @Test
@@ -99,13 +101,13 @@ class VireSimulationTest {
             }
         }
 
-        val component = simulation.createComponent(stateMachine)
+        val component = circuit.createComponent(stateMachine)
         val inputPort = component.createPort(inputVariable.variable())
         val outputPort = component.createPort(outputVariable.variable())
-        simulation.createNetworkNodeConnection(inputPort.node, outputPort.node)
+        circuit.createNetworkNodeConnection(inputPort.node, outputPort.node)
 
-        assertEquals(1, simulation.networks().size) { "More than one network remain after connecting the two nodes." }
-        val network = simulation.networks().first()
+        assertEquals(1, circuit.networks().size) { "More than one network remain after connecting the two nodes." }
+        val network = circuit.networks().first()
 
         assertEquals(network, inputPort.network) { "Input port network was not updated." }
         assertEquals(network, outputPort.network) { "Output port network was not updated." }
@@ -136,7 +138,7 @@ class VireSimulationTest {
             }
         }
 
-        val component = simulation.createComponent(stateMachine)
+        val component = circuit.createComponent(stateMachine)
         val outputPort = component.createPort(outputVariable.variable())
 
         // Simulate output.
@@ -148,12 +150,12 @@ class VireSimulationTest {
         assertEquals(outputState, node2.network.state)
 
         // Separate the nodes.
-        simulation.removeNetworkNodeConnection(node1, node2)
+        circuit.removeNetworkNodeConnection(node1, node2)
         assertEquals(outputState, node1.network.state)
         assertEquals(LogicState.EMPTY, node2.network.state)
 
         // Connect the nodes.
-        simulation.createNetworkNodeConnection(node1, node2)
+        circuit.createNetworkNodeConnection(node1, node2)
         assertEquals(outputState, node1.network.state)
         assertEquals(outputState, node2.network.state)
     }
